+++ src/agents/graphs/agent_graph_builder.py
@@ -14,9 +14,13 @@ from .graph_nodes import (
 from .interaction_handlers import (
     handle_ask_clarification_node,
     handle_continue_collaboration_node,
+    handle_create_project_node,  # noqa: F401 - imported for future use
     handle_deep_analysis_node,
     handle_idle_node,
+    handle_join_project_node,  # noqa: F401 - imported for future use
+    handle_leave_project_node,  # noqa: F401 - imported for future use
     handle_propose_idea_node,
+    handle_send_direct_message_node,  # noqa: F401 - imported for future use
 )
